Playing almost ok
