+++ jwarble/src/test/java/org/noise_planet/jwarble/OpenWarbleTest.java
@@ -17,6 +17,7 @@ import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.FutureTask;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
 
@@ -60,7 +61,7 @@ public class OpenWarbleTest {
         }
     }
 
-    private static void writeDoubleToFile(String path, double[] signal) throws IOException {
+    public static void writeDoubleToFile(String path, double[] signal) throws IOException {
         short[] shortSignal = new short[signal.length];
         double maxValue = Double.MIN_VALUE;
         for (double aSignal : signal) {
@@ -644,55 +645,4 @@ public class OpenWarbleTest {
             System.out.println(String.format("Odd %d Hz Even %d Hz", Math.round(openWarble.frequencies[idFreq]), Math.round(openWarble.frequenciesUptone[idFreq])));        }
     }
 
-    @Test
-    public void testMicrophone() throws Exception {
-        double sampleRate = 44100;
-        double powerPeak = 1; // 90 dBspl
-        double blankTime = 1.3;
-        int blankSamples = (int)(blankTime * sampleRate);
-        byte[] payload = new byte[] {18, 32, -117, -93, -50, 2, 52, 26, -117, 93, 119, -109, 39, 46, 108, 4, 31, 36, -100, 95, -9, -70, -82, -93, -75, -32, -63, 42, -44, -100, 50, 83, -118, 114};
-        OpenWarble openWarble = new OpenWarble(Configuration.getAudible(payload.length, sampleRate, true));
-        UtCallback utCallback = new UtCallback(false);
-        UtMessageCallback messageCallback = new UtMessageCallback();
-        openWarble.setCallback(messageCallback);
-        openWarble.setUnitTestCallback(utCallback);
-        double[] signal = openWarble.generateSignal(powerPeak, payload);
-        double[] allSignal = new double[blankSamples+signal.length+blankSamples];
-        System.arraycopy(signal, 0, allSignal, blankSamples, signal.length);
-        int cursor = 0;
-        while (cursor < allSignal.length) {
-            int len = Math.min(openWarble.getMaxPushSamplesLength(), allSignal.length - cursor);
-            if(len == 0) {
-                break;
-            }
-            openWarble.pushSamples(Arrays.copyOfRange(allSignal, cursor, cursor+len));
-            cursor+=len;
-        }
-        short[] shortSignal = new short[allSignal.length];
-        double maxValue = Double.MIN_VALUE;
-        for (double aSignal : allSignal) {
-            maxValue = Math.max(maxValue, aSignal);
-        }
-        maxValue *= 2;
-        for(int i=0; i<allSignal.length;i++) {
-            shortSignal[i] = (short)((allSignal[i] / maxValue) * Short.MAX_VALUE);
-        }
-        AtomicBoolean doRecord = new AtomicBoolean(true);
-        AtomicBoolean micOpen = new AtomicBoolean(false);
-        try {
-            FutureTask<double[]> recordTask = new FutureTask<>(new PlayRecord.Recorder(doRecord, micOpen));
-            ExecutorService executorService = Executors.newSingleThreadExecutor();
-            executorService.submit(recordTask);
-            while(!micOpen.get()) {
-                Thread.sleep(120);
-            }
-            PlayRecord playRecord = new PlayRecord();
-            playRecord.playSound(shortSignal, sampleRate);
-            doRecord.set(false);
-            double[] samples = recordTask.get();
-            executorService.shutdown();
-        } finally {
-            doRecord.set(false);
-        }
-    }
 }
